Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> (4 comments)

tried a bunch of combinations in Dev - http_proxy really wants a string, not an undef if it's not being used.  You are correct that ~ should not be quoted, puppet then rejects it as undef, not a string.

looks like the least pain path is to use '' as the fall-back default value - which works just fine.

Patch-set: 3
Reviewer: Gerrit User 1000237 <1000237@f182a8af-62ca-4ceb-ab81-14f8fec9c6c5>
Label: Verified=0